renamed utility folder to util to follow the convention throughout the app
